♻️ [Refactor]: chat 관련 component 적용
- MessageBox: speech balloon과 같은 UI
- UserMessageBox: User가 동일한 Date에 보낸 메세지들과 관련된 UI
- DateMessageBox: 동일한 날짜를 묶는 container UI
